Refactored 'amami/stash_utils.py' to be included inside 'amami/um_utils.py'.
